feat: fetch all users and show 9 cards at once
fixing the duplicate rendering of cards
